Files: *u launch.py - change import from file_names to file_management *u file_management.py - add method create_new_file_name from file_names.py *r file_names.py
